+++ modules/septima-lang/src/find-array-method.ts
@@ -97,6 +97,15 @@ export function findArrayMethod(arr: unknown[], index: string, callEvaluator: Ca
   if (index === 'some') {
     return Value.foreign(predicate => arr.some(adjustedPredicate(predicate)))
   }
+  if (index === 'sort') {
+    return Value.foreign(callback => {
+      if (callback) {
+        return [...arr].sort((a, b) => callEvaluator(callback, [Value.from(a), Value.from(b)]).assertNum())
+      } else {
+        return [...arr].sort()
+      }
+    })
+  }
   throw new Error(`Unrecognized array method: ${index}`)
 }